+++ tests/ttnn/unit_tests/operations/ccl/test_broadcast_op.py
@@ -21,7 +21,7 @@ def run_with_trace(
 ):
     # Compile Run
     logger.info("Compiling model")
-    tt_out_tensor = ttnn.experimental.broadcast(
+    tt_out_tensor = ttnn.broadcast(
         input_tensor_mesh,
         sender_coord=sender_coord,
         num_links=num_links,
@@ -35,7 +35,7 @@ def run_with_trace(
     logger.info("Capturing trace")
     trace_id = ttnn.begin_trace_capture(mesh_device, cq_id=0)
     for i in range(num_iter):
-        tt_out_tensor = ttnn.experimental.broadcast(
+        tt_out_tensor = ttnn.broadcast(
             input_tensor_mesh,
             sender_coord=sender_coord,
             num_links=num_links,
@@ -197,7 +197,7 @@ def run_broadcast_impl(
         tt_out_tensor_list.append(tt_out_tensor)
     else:
         for i in range(num_iters):
-            tt_out_tensors = ttnn.experimental.broadcast(
+            tt_out_tensors = ttnn.broadcast(
                 input_tensor_mesh_list[i],
                 sender_coord=sender_coord,
                 num_links=num_links,
