+++ ttnn/cpp/ttnn/operations/ccl/broadcast/device/kernels/broadcast_rm_writer.cpp
@@ -31,12 +31,13 @@ constexpr uint32_t num_rows_per_packet = get_compile_time_arg_val(4);
 constexpr uint32_t num_packets_per_row = get_compile_time_arg_val(5);
 constexpr uint32_t num_targets_forward_direction = get_compile_time_arg_val(6);
 constexpr uint32_t num_targets_backward_direction = get_compile_time_arg_val(7);
-constexpr uint32_t start_distance_in_hops_forward = get_compile_time_arg_val(8);
-constexpr uint32_t range_hops_forward = get_compile_time_arg_val(9);
-constexpr uint32_t start_distance_in_hops_backward = get_compile_time_arg_val(10);
-constexpr uint32_t range_hops_backward = get_compile_time_arg_val(11);
+constexpr uint32_t is_sender = get_compile_time_arg_val(8);
+constexpr uint32_t start_distance_in_hops_forward = get_compile_time_arg_val(9);
+constexpr uint32_t range_hops_forward = get_compile_time_arg_val(10);
+constexpr uint32_t start_distance_in_hops_backward = get_compile_time_arg_val(11);
+constexpr uint32_t range_hops_backward = get_compile_time_arg_val(12);
 
-inline constexpr uint32_t sharded_args_start_idx = 12;
+inline constexpr uint32_t sharded_args_start_idx = 13;
 
 /*
  * CCL Send will present various operating modes. Although there is only a single send kernel, it may (compile time)
@@ -134,92 +135,97 @@ void kernel_main() {
     noc_semaphore_set(reinterpret_cast<volatile tt_l1_ptr uint32_t*>(barrier_sem), 0);
 
     // 1. mcast via fabric to remote tensor addresses
-    uint32_t row_id = row_id_start;
-    while (row_id < row_id_end) {
-        size_t l1_read_addr = get_read_ptr(cb0_id);
-        cb_wait_front(cb0_id, num_rows_per_packet);
+    if (is_sender) {
+        uint32_t row_id = row_id_start;
+        while (row_id < row_id_end) {
+            size_t l1_read_addr = get_read_ptr(cb0_id);
+            cb_wait_front(cb0_id, num_rows_per_packet);
 
-        if constexpr (num_rows_per_packet == 1) {
-            uint32_t offset = 0;
-            uint32_t bytes_remaining = page_size;
-            for (uint32_t j = 0; j < num_packets_per_row && bytes_remaining > 0; j++) {
-                uint32_t packet_size = std::min(max_packet_size, bytes_remaining);
+            if constexpr (num_rows_per_packet == 1) {
+                uint32_t offset = 0;
+                uint32_t bytes_remaining = page_size;
+                for (uint32_t j = 0; j < num_packets_per_row && bytes_remaining > 0; j++) {
+                    uint32_t packet_size = std::min(max_packet_size, bytes_remaining);
 
-                noc_async_write(l1_read_addr + offset, tensor0_addrgen.get_noc_addr(row_id, offset), packet_size);
-                fabric_multicast_noc_unicast_write_with_state<
-                    UnicastWriteUpdateMask::DstAddr | UnicastWriteUpdateMask::PayloadSize>(
-                    fabric_connection,
-                    unicast_route_id,
-                    l1_read_addr + offset,
-                    tt::tt_fabric::NocUnicastCommandHeader{
-                        linear::addrgen_detail::get_noc_address(tensor0_addrgen, row_id, offset)},
-                    packet_size);
-                noc_async_writes_flushed();
-                offset += packet_size;
-                bytes_remaining -= packet_size;
-            }
-            row_id++;
-        } else {
-            uint32_t num_pages_for_current_packet = std::min<uint32_t>(row_id_end - row_id, num_rows_per_packet);
-            if (num_pages_for_current_packet == 1) {
-                noc_async_write(l1_read_addr, tensor0_addrgen.get_noc_addr(row_id, 0), page_size);
-                fabric_multicast_noc_unicast_write_with_state<UnicastWriteUpdateMask::DstAddr>(
-                    fabric_connection,
-                    unicast_route_id,
-                    l1_read_addr,
-                    tt::tt_fabric::NocUnicastCommandHeader{
-                        linear::addrgen_detail::get_noc_address(tensor0_addrgen, row_id, 0)},
-                    page_size);
-                noc_async_writes_flushed();
-                l1_read_addr += page_size;
+                    noc_async_write(l1_read_addr + offset, tensor0_addrgen.get_noc_addr(row_id, offset), packet_size);
+                    fabric_multicast_noc_unicast_write_with_state<
+                        UnicastWriteUpdateMask::DstAddr | UnicastWriteUpdateMask::PayloadSize>(
+                        fabric_connection,
+                        unicast_route_id,
+                        l1_read_addr + offset,
+                        tt::tt_fabric::NocUnicastCommandHeader{
+                            linear::addrgen_detail::get_noc_address(tensor0_addrgen, row_id, offset)},
+                        packet_size);
+                    noc_async_writes_flushed();
+                    offset += packet_size;
+                    bytes_remaining -= packet_size;
+                }
                 row_id++;
-            } else if (num_pages_for_current_packet == 2) {
-                noc_async_write(l1_read_addr, tensor0_addrgen.get_noc_addr(row_id, 0), page_size);
-                noc_async_write(l1_read_addr + page_size, tensor0_addrgen.get_noc_addr(row_id + 1, 0), page_size);
-                fabric_multicast_noc_scatter_write_with_state<UnicastScatterWriteUpdateMask::DstAddrs>(
-                    fabric_connection,
-                    scatter_route_id,
-                    l1_read_addr,
-                    tt::tt_fabric::NocUnicastScatterCommandHeader{
-                        {linear::addrgen_detail::get_noc_address(tensor0_addrgen, row_id, 0),
-                         linear::addrgen_detail::get_noc_address(tensor0_addrgen, row_id + 1, 0)},
-                        static_cast<uint16_t>(page_size)},  // ignore
-                    page_size * 2);                         // ignore
-                noc_async_writes_flushed();
-                l1_read_addr += page_size * 2;
-                row_id += 2;
             } else {
-                ASSERT(false);
+                uint32_t num_pages_for_current_packet = std::min<uint32_t>(row_id_end - row_id, num_rows_per_packet);
+                if (num_pages_for_current_packet == 1) {
+                    noc_async_write(l1_read_addr, tensor0_addrgen.get_noc_addr(row_id, 0), page_size);
+                    fabric_multicast_noc_unicast_write_with_state<UnicastWriteUpdateMask::DstAddr>(
+                        fabric_connection,
+                        unicast_route_id,
+                        l1_read_addr,
+                        tt::tt_fabric::NocUnicastCommandHeader{
+                            linear::addrgen_detail::get_noc_address(tensor0_addrgen, row_id, 0)},
+                        page_size);
+                    noc_async_writes_flushed();
+                    l1_read_addr += page_size;
+                    row_id++;
+                } else if (num_pages_for_current_packet == 2) {
+                    noc_async_write(l1_read_addr, tensor0_addrgen.get_noc_addr(row_id, 0), page_size);
+                    noc_async_write(l1_read_addr + page_size, tensor0_addrgen.get_noc_addr(row_id + 1, 0), page_size);
+                    fabric_multicast_noc_scatter_write_with_state<UnicastScatterWriteUpdateMask::DstAddrs>(
+                        fabric_connection,
+                        scatter_route_id,
+                        l1_read_addr,
+                        tt::tt_fabric::NocUnicastScatterCommandHeader{
+                            {linear::addrgen_detail::get_noc_address(tensor0_addrgen, row_id, 0),
+                             linear::addrgen_detail::get_noc_address(tensor0_addrgen, row_id + 1, 0)},
+                            static_cast<uint16_t>(page_size)},  // ignore
+                        page_size * 2);                         // ignore
+                    noc_async_writes_flushed();
+                    l1_read_addr += page_size * 2;
+                    row_id += 2;
+                } else {
+                    ASSERT(false);
+                }
             }
+            cb_pop_front(cb0_id, num_rows_per_packet);
         }
-        cb_pop_front(cb0_id, num_rows_per_packet);
+
+        // 2. mcast output ready semaphore
+        uint64_t out_ready_sem_noc_addr_in_pkt =
+            safe_get_noc_addr(out_ready_sem_noc0_x, out_ready_sem_noc0_y, out_ready_sem_bank_addr, 0);
+
+        fabric_multicast_noc_unicast_atomic_inc_with_state<UnicastAtomicIncUpdateMask::DstAddr>(
+            fabric_connection,
+            sem_route_id,
+            tt::tt_fabric::NocUnicastAtomicIncCommandHeader{out_ready_sem_noc_addr_in_pkt, 0, 0});
+        // increment locally
+        uint64_t out_ready_sem_noc_addr =
+            safe_get_noc_addr(out_ready_sem_noc0_x, out_ready_sem_noc0_y, out_ready_sem_bank_addr);
+        noc_semaphore_inc(out_ready_sem_noc_addr, 1);
+
+        close_connections(fabric_connection);
+
+        noc_async_write_barrier();
+    } else {
+        if (wait_output_semaphore) {
+            volatile tt_l1_ptr uint32_t* sem_ptr =
+                reinterpret_cast<volatile tt_l1_ptr uint32_t*>(out_ready_sem_bank_addr);
+            noc_semaphore_wait(sem_ptr, out_ready_sem_wait_value);
+        }
+
+        // 4. global semaphore reset
+        if (reset_global_semaphore) {
+            noc_semaphore_set(reinterpret_cast<volatile tt_l1_ptr uint32_t*>(out_ready_sem_bank_addr), 0);
+        }
+        close_connections(fabric_connection);
+
+        noc_async_write_barrier();
     }
-
-    // 2. mcast output ready semaphore
-    uint64_t out_ready_sem_noc_addr_in_pkt =
-        safe_get_noc_addr(out_ready_sem_noc0_x, out_ready_sem_noc0_y, out_ready_sem_bank_addr, 0);
-
-    fabric_multicast_noc_unicast_atomic_inc_with_state<UnicastAtomicIncUpdateMask::DstAddr>(
-        fabric_connection,
-        sem_route_id,
-        tt::tt_fabric::NocUnicastAtomicIncCommandHeader{out_ready_sem_noc_addr_in_pkt, 0, 0});
-    // increment locally
-    uint64_t out_ready_sem_noc_addr =
-        safe_get_noc_addr(out_ready_sem_noc0_x, out_ready_sem_noc0_y, out_ready_sem_bank_addr);
-    noc_semaphore_inc(out_ready_sem_noc_addr, 1);
-
-    // 3. wait for mcast output ready semaphore
-    if (wait_output_semaphore) {
-        volatile tt_l1_ptr uint32_t* sem_ptr = reinterpret_cast<volatile tt_l1_ptr uint32_t*>(out_ready_sem_bank_addr);
-        noc_semaphore_wait(sem_ptr, out_ready_sem_wait_value);
-    }
-
-    // 4. global semaphore reset
-    if (reset_global_semaphore) {
-        noc_semaphore_set(reinterpret_cast<volatile tt_l1_ptr uint32_t*>(out_ready_sem_bank_addr), 0);
-    }
-
-    close_connections(fabric_connection);
-
-    noc_async_write_barrier();
 }
