+++ ttnn/cpp/ttnn/operations/ccl/ccl_pybind.cpp
@@ -11,6 +11,7 @@
 #include "ttnn/operations/ccl/barrier/barrier_pybind.hpp"
 #include "ttnn/operations/ccl/all_gather/all_gather_pybind.hpp"
 #include "ttnn/operations/ccl/all_to_all_combine/all_to_all_combine_pybind.hpp"
+#include "ttnn/operations/ccl/broadcast/broadcast_pybind.hpp"
 #include "ttnn/operations/ccl/all_to_all_dispatch/all_to_all_dispatch_pybind.hpp"
 #include "ttnn/operations/ccl/reduce_scatter/reduce_scatter_pybind.hpp"
 #include "ttnn/operations/ccl/all_reduce/all_reduce_pybind.hpp"
@@ -35,6 +36,7 @@ void py_module(py::module& module) {
     ccl::py_bind_all_to_all_dispatch(module);
     ccl::py_bind_reduce_scatter(module);
     ccl::py_bind_all_reduce(module);
+    ccl::py_bind_broadcast(module);
 }
 
 }  // namespace ttnn::operations::ccl
